+++ taf/auth_repo.py
@@ -66,7 +66,6 @@ class AuthenticationRepository(GitRepository, TAFRepository):
             *args,
             **kwargs,
         )
-
         if conf_directory_root is None:
             conf_directory_root_path = Path(self.path).parent
         else: