+++ taf/updater/updater_pipeline.py
@@ -17,6 +17,7 @@ import taf.settings as settings
 import taf.repositoriesdb as repositoriesdb
 from taf.auth_repo import AuthenticationRepository
 from taf.exceptions import (
+    InvalidRepositoryError,
     MissingInfoJsonError,
     RepositoryNotCleanError,
     UpdateFailedError,
@@ -201,6 +202,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         out_of_band_authentication,
         checkout,
         excluded_target_globs,
+        bare,
     ):
 
         super().__init__(
@@ -248,6 +250,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
         self.out_of_band_authentication = out_of_band_authentication
         self.checkout = checkout
         self.excluded_target_globs = excluded_target_globs
+        self.bare = bare
         self.state = UpdateState()
         self.state.targets_data = {}
         self._output = None
@@ -360,7 +363,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             # Use ThreadPoolExecutor to run _clone_validation_repo in a separate thread
             with ThreadPoolExecutor() as executor:
                 future_validation_repo = executor.submit(
-                    _clone_validation_repo, self.url
+                    _clone_validation_repo, self.url, True
                 )
                 validation_repo = future_validation_repo.result()
 
@@ -382,8 +385,9 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 auth_repo_name = _get_repository_name_raise_error_if_not_defined(
                     validation_repo, top_commit_of_validation_repo
                 )
-
-            git_updater = GitUpdater(self.url, self.library_dir, validation_repo.name)
+            git_updater = GitUpdater(
+                self.url, self.library_dir, validation_repo.name, self.bare
+            )
             last_validated_remote_commit, error = _run_tuf_updater(
                 git_updater, auth_repo_name
             )
@@ -401,21 +405,18 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     )
                 else:
                     self.state.auth_repo_name = auth_repo_name
-
             self.state.users_auth_repo = AuthenticationRepository(
-                self.library_dir,
-                self.state.auth_repo_name,
+                path=self.auth_path,
+                library_dir=self.library_dir,
+                name=self.state.auth_repo_name,
                 urls=[self.url],
             )
-
             self.state.existing_repo = self.state.users_auth_repo.is_git_repository_root
             self._validate_operation_type()
             self.state.validation_auth_repo = git_updater.validation_auth_repo
             self.state.is_test_repo = self.state.validation_auth_repo.is_test_repo
-
             if self.operation == OperationType.UPDATE:
                 self._validate_last_validated_commit(settings.last_validated_commit)
-
             # used for testing purposes
             if settings.overwrite_last_validated_commit:
                 self.state.last_validated_commit = settings.last_validated_commit
@@ -423,7 +424,6 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                 self.state.last_validated_commit = (
                     self.state.users_auth_repo.last_validated_commit
                 )
-
             if error is None:
                 self.state.auth_commits_since_last_validated = list(git_updater.commits)
                 taf_logger.info(
@@ -553,7 +553,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
             if self.state.existing_repo:
                 self.state.users_auth_repo.fetch(fetch_all=True)
             else:
-                self.state.users_auth_repo.clone()
+                self.state.users_auth_repo.clone(bare=self.bare)
         except Exception as e:
             self.state.errors.append(e)
             self.state.event = Event.FAILED
@@ -637,7 +637,7 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
                     )
                     self.state.repos_on_disk[users_repo.name] = users_repo
                 else:
-                    temp_repo.clone(bare=True)
+                    temp_repo.clone(bare=self.bare)
                     self.state.repos_not_on_disk[users_repo.name] = users_repo
 
             with ThreadPoolExecutor() as executor:
@@ -865,19 +865,30 @@ class AuthenticationRepositoryUpdatePipeline(Pipeline):
 
     @log_on_start(
         DEBUG,
-        "Checking if target repositories contain unpushed commits...",
+        "Checking if target repositories are clean...",
         logger=taf_logger,
     )
     def check_if_local_target_repositories_clean(self):
         try:
             for repository in self.state.repos_on_disk.values():
-                if repository.something_to_commit():
-                    raise RepositoryNotCleanError(repository.name)
-                for branch in self.state.target_branches_data_from_auth_repo[
-                    repository.name
-                ]:
-                    if repository.is_branch_with_unpushed_commits(branch):
-                        raise UnpushedCommitsError(repository.name, branch)
+                if repository.is_bare_repository:
+                    # For bare repositories, ensure they are valid
+                    if not repository.is_git_repository:
+                        raise InvalidRepositoryError(
+                            f"{repository.name} is not a valid git repository."
+                        )
+                    taf_logger.debug(
+                        f"Repo {repository.name} is a bare repository and is valid."
+                    )
+                else:
+                    # For non-bare repositories, check for uncommitted changes and unpushed commits
+                    if repository.something_to_commit():
+                        raise RepositoryNotCleanError(repository.name)
+                    for branch in self.state.target_branches_data_from_auth_repo[
+                        repository.name
+                    ]:
+                        if repository.is_branch_with_unpushed_commits(branch):
+                            raise UnpushedCommitsError(repository.name, branch)
         except Exception as e:
             self.state.errors.append(e)
             self.state.event = Event.FAILED
@@ -1110,9 +1121,10 @@ but commit not on branch {current_branch}"
                 ]
                 temp_target_repo = self.state.temp_target_repositories[repository_name]
                 users_target_repo.clone_from_disk(
-                    temp_target_repo.path, temp_target_repo.get_remote_url()
+                    temp_target_repo.path,
+                    temp_target_repo.get_remote_url(),
+                    is_bare=self.bare,
                 )
-
             for repo_name in self.state.repos_on_disk:
                 users_target_repo = self.state.users_target_repositories[repo_name]
                 temp_target_repo = self.state.temp_target_repositories[repo_name]
@@ -1150,6 +1162,9 @@ but commit not on branch {current_branch}"
         """Determines which commits needs to be merged into the specified branch and
         merge it.
         """
+        if self.bare:
+            taf_logger.info("Skipping merge_commits for bare repository")
+            return UpdateStatus.SUCCESS
         try:
             if self.only_validate:
                 return self.state.update_status
@@ -1283,7 +1298,7 @@ but commit not on branch {current_branch}"
                     )
 
 
-def _clone_validation_repo(url):
+def _clone_validation_repo(url, bare=False):
     """
     Clones the authentication repository based on the url specified using the
     mirrors parameter. The repository is cloned as a bare repository
@@ -1464,7 +1479,8 @@ def _merge_commit(repository, branch, commit_to_merge, force_revert=True):
     """Merge the specified commit into the given branch and check out the branch.
     If the repository cannot contain unauthenticated commits, check out the merged commit.
     """
-
+    if repository.is_bare_repository:
+        return
     try:
         repository.checkout_branch(branch, raise_anyway=True)
     except GitError as e: